New podman::quadlet type for quadlet units
Assuming at least podman 4.4.0 the new type `podman::quadlet`
will create a quadlet unit files for a container, volume or pod
which will in turn generate a service unit for managing the container,
pod or volume.

Boot root and rootless containers are supported.

The included examples are operational however.

Note that pod quadlets are only supported with podman 5.